Refactor Praser checkUserInput method
